Remove references to internal signing service
The scripts were never in GitHub, but my Mercurial `.hgrc` config
referred to a python script used at work that was used to sign
commits. This was pretty pointless since I only use the script at
work, so have moved the necessary config to `/etc/mercurial/hgrc`.

Hopefully I'll manage to find that in the future, or happen upon
this commit message...